+++ CHANGELOG.md
@@ -4,10 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2021-12-29
+### Added
+- Volume control for SoundPad (Tabletop)
+- SoundPad volume stored as world settings
+
 ## [3.0.1] - 2021-12-23
+### Changed
 - Add support for FVTT 9.x version
 
 ## [2.13.0] - 2021-12-12
+### Added
 - Add button to open SoundPad UI without the shortcut (ALT+S)
 
 ## [2.12.0] - 2021-12-04